stabilize worker total busy duration, bring WorkerMetrics, MetricsBatch and Histogram out of unstable flag
